pyproject.toml
[project]
name = "hri"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.65.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
